Updates by Bozingle
The constructor either loads an existing processed data file or processes a new JR_Data object that is saved to the current file that the program is located in. If this feature needs to be changed, the lines 16, 17, and 26 need to be changed to the correct file location. The JR_Data object contains 2 tall arrays called audio and spgram.