Update(1/9/2019)
This update runs through the audio data and saves the intervals of the spectrogram into memory. Current issue: The amount of indexes produced is equivalent to the entire spectrogram, but does not read that 15 minutes has been elapsed. This runs through very quickly. The idea is to load the intervals that should be displayed with the "get" function that will be made in the future.