Update
obj.audio is now a tall array, and no longer obj.audio.audio. obj.spgram is now the tall array of the spectrogram. You can receive the intervals by the method demonstrated in test.m. There is now a propertyType so the get function is flexible with either retrieving audio or spectrograms and any interval. There's an error case when an incorrect propertyType is given.